Add files via upload
Hey Guys,

 I Worked on a generic interface for Team, Model, Controller(Sorting, TryUpdateModel, and Filtering), and View. I also Incorporated an Elephant Controller that remembers each page name alongside the WMBA Logo attached to it. Lastly, I worked on paging. 

The Create View for teams needs to be worked on properly, cause I couldn't debug it properly cause of time and, the players is not being retrieved appropriately yet. 

Ps: We need to figure how we can give a descriptive Info on the Game Teams row. I.e: Blue Jays VS Welland Jackfish.